+++ changelog.md
@@ -9,4 +9,5 @@
 
 ## 1.2.0 [WIP]
 - Configured [forever](https://www.npmjs.com/package/forever) log files. 
-- Keep already built Highcharts files for improved performance. 
+- Keep already built Highcharts files for improved performance.
+- Added functionality to validate a Webhook from Github.